// frontend/user/app/types/merger.ts
export interface MergerUnit {
    id: number;
    matinh: number;
    ma: string;
    tentinh: string;
    loai: string;
    tenhc: string;
    cay: string;
    dientichkm2: number;
    dansonguoi: string;
    trungtamhc: string;
    kinhdo: number;
    vido: number;
    truocsapnhap: string;
    maxa: number;
    khoa: string;
}

export interface ProvinceListItem {
    mahc: string;
    tentinh: string;
}